Move initialization systems to their relevant modules
